third_party/WebKit/LayoutTests/platform/linux/virtual/mojo-blobs/external/wpt/FileAPI/blob/Blob-XHR-revoke-expected.txt

This is a testharness.js-based test.
FAIL Revoke blob URL after open(), will fetchassert_equals: expected "test" but got ""
FAIL Revoke blob URL before open(), network error (after send())assert_unreached: Got unexpected load event Reached unreachable code
Harness: the test ran to completion.
